refactor(build): move platform abstraction files into platform/ subdirectory
Keeps main directory focused on core spine sources; groups the POSIX/Windows
abstraction layer together per upstream review feedback.
