+++ nft_popen.c
@@ -86,8 +86,8 @@
 
 #include "common.h"
 #include "spine.h"
-#include "platform_error.h"
-#include "platform_process.h"
+#include "platform/platform_error.h"
+#include "platform/platform_process.h"
 #include <spawn.h>
 
 /* An instance of this struct is created for each popen() fd. */
