+++ platform/platform.h
