+++ platform/platform_common.c
