+++ platform/platform_error.h
