+++ platform/platform_error_posix.c
