+++ platform/platform_error_win.c
