+++ platform/platform_fd.h
