+++ platform/platform_fd_posix.c
