+++ platform/platform_fd_win.c
