+++ platform/platform_posix.c
