+++ platform/platform_socket_win.c
