+++ platform/platform_win.c
